requirements.txt
exceptiongroup==1.3.0
iniconfig==2.1.0
numpy==2.2.6
numpy-quaternion==2024.0.8
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
scipy==1.15.3
spiceypy==6.0.0
tomli==2.2.1
typing_extensions==4.13.2
trimesh
pyyaml